+++ CHANGELOG.md
@@ -194,6 +194,7 @@ This version of std++ newly supports Rocq 9.1.
 - Add lemma `rtsc_nf`.
 - Add instance `Symmetric R → Equivalence (rtc R)`. Previously this was a lemma
   called `rtc_equivalence`.
+- Add `Countable` instance for `coPset`. (by Benjamin Peters)
 
 **Changes in `stdpp_bitvector`:**